+++ animation_nodes/nodes/material/object_material_output.py
@@ -1,9 +1,13 @@
 import bpy
+from bpy.props import *
 from ... base_types import AnimationNode, VectorizedSocket
 
 class ObjectMaterialOutputNode(bpy.types.Node, AnimationNode):
     bl_idname = "an_ObjectMaterialOutputNode"
     bl_label = "Object Material Output"
+
+    appendMaterials: BoolProperty(name = "Append Materials", default = False,
+        description = "Append input material(s) to the object's materials instead of overwriting them")
 
     useMaterialList: VectorizedSocket.newProperty()
 
@@ -15,6 +19,10 @@ class ObjectMaterialOutputNode(bpy.types.Node, AnimationNode):
 
         self.newOutput("Object", "Object", "object")
 
+    def drawAdvanced(self, layout):
+        row = layout.row(align = True)
+        row.prop(self, "appendMaterials")
+
     def getExecutionFunctionName(self):
         if self.useMaterialList:
             return "executeList"
@@ -23,13 +31,25 @@ class ObjectMaterialOutputNode(bpy.types.Node, AnimationNode):
 
     def executeSingle(self, object, material):
         if object is None or not hasattr(object.data, "materials"): return object
-        object.data.materials.clear()
-        object.data.materials.append(material)
+
+        objectMaterials = object.data.materials
+        if not self.appendMaterials: objectMaterials.clear()
+
+        self.appendMaterial(objectMaterials, material)
         return object
 
     def executeList(self, object, materials):
         if object is None or not hasattr(object.data, "materials"): return object
-        object.data.materials.clear()
+
+        objectMaterials = object.data.materials
+        if not self.appendMaterials: objectMaterials.clear()
+
         for material in materials:
-            object.data.materials.append(material)
+            self.appendMaterial(objectMaterials, material)
         return object
+
+    def appendMaterial(self, objectMaterials, material):
+        if self.appendMaterials:
+            if material.name not in objectMaterials: objectMaterials.append(material)
+        else:
+            objectMaterials.append(material)